+++ packages/database/src/index.js
@@ -4,3 +4,10 @@ export {
   saveTechnicalIndicators,
   createMarketDataRepository,
 } from "./marketDataRepository.js";
+export {
+  saveLlmRequest,
+  saveDecision,
+  saveOrder,
+  saveFill,
+  createTradingRepository,
+} from "./tradingRepository.js";